Move shape_defs to mesaac_measures.
