+++ src/lib/mesaac_common/include/mesaac_common/mesaac_common.hpp
@@ -3,4 +3,3 @@
 #include "b32.hpp"
 #include "b64.hpp"
 #include "gzip.hpp"
-#include "shape_defs.hpp"
