+++ src/lib/mesaac_measures/include/mesaac_measures/measures_base.hpp
@@ -7,7 +7,7 @@
 #include <memory>
 #include <string>
 
-#include "mesaac_common/shape_defs.hpp"
+#include "mesaac_measures/shape_defs.hpp"
 
 namespace mesaac::measures {
