+++ src/lib/mesaac_measures/include/mesaac_measures/mesaac_measures.hpp
@@ -6,6 +6,7 @@
 #include "hamann.hpp"
 #include "measures_base.hpp"
 #include "measures_factory.hpp"
+#include "shape_defs.hpp"
 #include "shape_measures_factory.hpp"
 #include "tanimoto.hpp"
 #include "tversky.hpp"
