+++ src/lib/mesaac_measures/src/shape_defs.cpp
@@ -1,4 +1,4 @@
-#include "mesaac_common/shape_defs.hpp"
+#include "mesaac_measures/shape_defs.hpp"
 
 namespace mesaac::shape_defs {
 BitVector bit_vector_from_str(const std::string &strval) {
