+++ src/lib/mesaac_shape/CMakeLists.txt
@@ -17,9 +17,7 @@ target_sources(
          include/mesaac_shape/vol_box.hpp)
 
 target_compile_features(${TARGET} PUBLIC cxx_std_20)
-# mesaac_common is needed only for shape_defs.hpp.  Should that header live
-# elsewhere?
-target_link_libraries(${TARGET} PUBLIC mesaac_common mesaac_mol svd ap
+target_link_libraries(${TARGET} PUBLIC mesaac_measures mesaac_mol svd ap
                                        Boost::dynamic_bitset)
 
 if(PROVIDE_EIGEN)
