+++ src/lib/mesaac_shape/include/mesaac_shape/fingerprinter.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "mesaac_common/shape_defs.hpp"
+#include "mesaac_measures/shape_defs.hpp"
 #include "mesaac_mol/atom.hpp"
 #include "mesaac_shape/vol_box.hpp"
