+++ src/lib/mesaac_shape/include/mesaac_shape/shared_types.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "mesaac_common/shape_defs.hpp"
+#include "mesaac_measures/shape_defs.hpp"
 #include <vector>
 
 namespace mesaac::shape {
